refactor(menus): split MenusPage.tsx into smaller files
`MenusPage.tsx` had grown to 2265 lines mixing the page shell, the
~930-line per-menu editor, rail list items, menu-generation logic, and
project-graph helpers in one file (#48). Before splitting, added
`MenusPage.test.tsx` (14 tests) as a characterization suite covering rail
selection, add/remove menu, all three menu generators, canvas zoom, view
toggles, and the empty-workspace state, run green against the pre-split
code first.

Split into:
- `components/menus/MenuEditor.tsx` — the unified per-menu editor (toolbar,
  canvas + tool palette, inspector, map view)
- `components/menus/MenuListItem.tsx` — rail entries and the empty-workspace
  panel
- `components/menus/menuGenerators.ts` — the chapter-grid/audio-setup/
  subtitle-setup pure menu builders
- `components/menus/menuProjectHelpers.ts` — menu lookup/update and the
  navigation-graph connection-count calculator
- `components/menus/menuDefaults.ts` — shared canvas/button/text defaults

`MenusPage.tsx` itself is now ~580 lines: just `MenusPage` and
`MenusWorkspace`. Updated `SceneEditor.test.tsx`'s import path for the
three generator functions, which moved out of `MenusPage.tsx`.

Verified with `tsc --noEmit`, `pnpm test:js` (85 passing, including the new
characterization suite), `pnpm format:check`, `pnpm build`, and manually in
the running app via the Tauri MCP bridge (seeded the project store directly,
navigated to Menus, confirmed rail connection counts and the Chapter Grid
generator both work against the split code).
